requirements.txt
flask
ultralytics
opencv-python
numpy
Pillow
streamlit
gtts
playsound
pyttsx3
pydub
torch
torchaudio
torchvision
requests
matplotlib
pandas
